@@ -1,4 +1,4 @@
 'use strict'
 // auto-generated by boundation, do not update manually
-/** @type {typeof import("null") } */
+/** @type {typeof import("./edition-types/test.d.ts") } */
 module.exports = require('editions').requirePackage(__dirname, require, 'test.js')
